ARM32 bit fixes, for 64bit printf format specifier (#490)
* ARM32 bit fixes, for 64bit printf format specifier


* [Warning] -Wlong-long is valid for ULL const in Arm32bit
pointer cast align warnings


* Warning fixes for using ULL and pointer type casting in 32 bit arch


* deserialized value didnt match, reverting the warning fix


* Fix for deserialized value not matching after warning changes


* Put back the reverted fix

+++ meta/saiserialize.cpp
@@ -2111,7 +2111,7 @@ void sai_deserialize_pointer(
 {
     SWSS_LOG_ENTER();
 
-    sai_deserialize_number(s, (uint64_t&)ptr, true);
+    sai_deserialize_number(s, (uintptr_t &)ptr, true);
 }
 
 void sai_deserialize_ip_address(
